support for ansible source
