+++ docs/user/index.rst
@@ -1,1 +1,2 @@
-Codev
+codev settings file:
+
